pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ws-foundational-training"
version = "0.1.0"
description = "Interactive educational program for Watts-Strogatz multi-modal neural architectures"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Multi-Modal WS Training Team"}
]
keywords = [
    "machine-learning",
    "neural-networks",
    "watts-strogatz",
    "small-world-networks",
    "sparse-training",
    "multi-modal",
    "education"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "networkx>=3.2",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.1.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.6.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]
full = [
    "torchaudio>=2.1.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.13.0",
]

[project.urls]
Homepage = "https://github.com/SamuraiBuddha/multi-modal-WS-foundational-training"
Documentation = "https://github.com/SamuraiBuddha/multi-modal-WS-foundational-training/tree/main/docs"
Repository = "https://github.com/SamuraiBuddha/multi-modal-WS-foundational-training"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
